chore: DoughnutChart 컴포넌트 로직 수정
